# src/api/api.yaml
openapi: 3.0.0
paths:
...
  /auth/local/signin:
    post:
      operationId: AuthController_signIn
      summary: Sign in using email and password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInDto'
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensDto'
      tags: &ref_2
        - Auth
  /auth/logout:
    post:
      operationId: AuthController_logout
      summary: Log out user and invalidate session
      parameters: []
      responses:
        '200':
          description: User successfully logged out
      tags: *ref_2
      security:
        - bearer: []
  /auth/reset-password-request:
    post:
      operationId: AuthController_resetPasswordRequest
      summary: Initiate password reset request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestResetPasswordDto'
      responses:
        '200':
          description: Reset password email sent (if user exists)
      tags: *ref_2
  /auth/admin/reset-password/{userId}:
    post:
      operationId: AuthController_adminResetPassword
      summary: 'Owner/admin: send a password reset email to a specific user'
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Password reset email sent to the user
      tags: *ref_2
      security:
        - bearer: []
  /auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Reset password using reset token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset successful
      tags: *ref_2
      security:
        - JWT-reset-password: []
  /auth/refresh:
    post:
      operationId: AuthController_refreshToken
      summary: Refresh access and refresh tokens
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensDto'
      tags: *ref_2
      security:
        - JWT-refresh: []
  /auth/register:
    post:
      operationId: AuthController_register
      summary: Finish registration by setting password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '200':
          description: User registered successfully
      tags: *ref_2
      security:
        - JWT-register-user: []
  /auth/check-register:
    post:
      operationId: AuthController_checkRegister
      summary: Check if registration token is valid
      parameters: []
      responses:
        '200':
          description: Returns email and first name if valid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  firstName:
                    type: string
      tags: *ref_2
      security:
        - JWT-register-user: []
...
  /image/list:
    get:
      operationId: ImageController_getList
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: List of image data with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageListResponseDto'
      tags: &ref_5
        - Image
  /image:
    get:
      operationId: ImageController_get
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Get single image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDataResponseDto'
      tags: *ref_5
    put:
      operationId: ImageController_update
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Update image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDataResponseDto'
      tags: *ref_5
      security:
        - bearer: []
    delete:
      operationId: ImageController_delete
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Delete image
      tags: *ref_5
      security:
        - bearer: []
  /image/create:
    post:
      operationId: ImageController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageDataDto'
      responses:
        '200':
          description: Create image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDataResponseDto'
      tags: *ref_5
      security:
        - bearer: []
  /image/cover:
    get:
      operationId: ImageController_getCoverImage
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Get image cover file stream
          content:
            image/*:
              schema:
                type: string
                format: binary
      tags: *ref_5
  /image/low-res:
    get:
      operationId: ImageController_getLowResImage
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Get low resolution image stream
          content:
            image/*:
              schema:
                type: string
                format: binary
      tags: *ref_5
  /image/original:
    get:
      operationId: ImageController_getOriginalImage
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Get original image file stream
          content:
            image/*:
              schema:
                type: string
                format: binary
      tags: *ref_5
      security:
        - bearer: []
...
  /dev/ocr/recognize:
    post:
      operationId: OcrController_recognize
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OcrOptionsDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrResponseDto'
      tags: &ref_23
        - OCR (dev)
      security: &ref_24
        - bearer: []
  /dev/ocr/structured:
    post:
      operationId: OcrController_structured
      summary: Recognize, clean, and structure OCR into items and totals
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrStructuredResponseDto'
      tags: *ref_23
      security: *ref_24
...
info:
  title: WHCP Backend
  description: XD
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
    JWT-refresh:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT refresh
      description: Enter JWT refresh token
      in: header
    JWT-reset-password:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT reset password
      description: Enter JWT reset password token
      in: header
    JWT-register-user:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT register user
      description: Enter JWT register user
      in: header
  schemas:
...
    PaginationDto:
      type: object
      properties:
        take:
          type: number
          minimum: 1
          maximum: 20
        skip:
          type: number
          minimum: 0
          default: 0
...
    SignInDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    TokensDto:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
      required:
        - access_token
        - refresh_token
    RequestResetPasswordDto:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        newPassword:
          type: string
        deleteSessions:
          type: boolean
      required:
        - newPassword
        - deleteSessions
    RegisterDto:
      type: object
      properties:
        password:
          type: string
      required:
        - password
...
    AuthorDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
    ImageDataResponseDto:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: '#/components/schemas/AuthorDto'
        dateTaken:
          type: string
          format: date-time
        imageId:
          type: string
        localization:
          type: string
        description:
          type: string
        title:
          type: string
      required:
        - id
        - author
        - dateTaken
        - imageId
    ImageListResponseDto:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/ImageDataResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - images
        - total
    ImageDataDto:
      type: object
      properties:
        localization:
          type: string
        dateTaken:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        authorId:
          type: string
        imageId:
          type: string
      required:
        - localization
        - dateTaken
        - imageId
...
    OcrLang:
      type: string
      default: pol
      enum:
        - pol
        - eng
        - deu
        - spa
        - fra
    OcrPSM:
      type: number
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        - 13
    OcrOEM:
      type: number
      enum:
        - 0
        - 1
        - 2
        - 3
    OcrOptionsDto:
      type: object
      properties:
        lang:
          $ref: '#/components/schemas/OcrLang'
        psm:
          $ref: '#/components/schemas/OcrPSM'
        oem:
          $ref: '#/components/schemas/OcrOEM'
        rotate:
          type: integer
          default: 0
        maxWidth:
          type: integer
          default: 1800
          description: Max image width after resize
          minimum: 200
          maximum: 4000
        timeoutMs:
          type: integer
          default: 15000
        threshold:
          type: boolean
          default: false
      required:
        - rotate
        - timeoutMs
    OcrUsedDto:
      type: object
      properties:
        lang:
          type: string
        psm:
          $ref: '#/components/schemas/OcrPSM'
        oem:
          $ref: '#/components/schemas/OcrOEM'
        rotate:
          type: number
        width:
          type: number
        threshold:
          type: boolean
      required:
        - lang
        - psm
        - oem
        - rotate
        - width
        - threshold
    OcrResponseDto:
      type: object
      properties:
        text:
          type: string
        durationMs:
          type: number
        used:
          $ref: '#/components/schemas/OcrUsedDto'
        shotDate:
          type: string
      required:
        - text
        - durationMs
        - used
    OcrStructuredHeaderDto:
      type: object
      properties:
        sellerName:
          type: string
          description: Seller/vendor name
          example: APPLOGY
        sellerTaxId:
          type: string
          description: Seller tax id (e.g., NIP/VATIN)
          example: '1122009977'
        date:
          type: string
          description: Document date (ISO-8601 if possible)
          example: '2018-08-14'
        documentNumber:
          type: string
          description: Document number (invoice no., receipt id, etc.)
          example: H000156
        buyerName:
          type: string
          description: Buyer name if present
          example: ACME Sp. z o.o.
        buyerTaxId:
          type: string
          description: Buyer tax id if present
          example: '1234567890'
    OcrStructuredItemDto:
      type: object
      properties:
        name:
          type: string
          description: Item name as seen on the document
          example: Apples
        quantity:
          type: number
          description: Quantity (if available)
          example: 2
        unit:
          type: string
          description: Unit of measure (e.g., kg, pcs)
          example: kg
        unitPrice:
          type: number
          description: Unit price (if available)
          example: 7.5
        lineTotal:
          type: number
          description: Line total (sum for this item)
          example: 15
        vatRate:
          type: number
          description: VAT rate in percent (if present)
          example: 8
        currency:
          type: string
          description: Currency code inferred from the document
          example: PLN
      required:
        - name
    OcrStructuredTotalsDto:
      type: object
      properties:
        subtotal:
          type: number
          description: Subtotal amount before taxes (if present)
          example: 32.37
        vatTotal:
          type: number
          description: Total VAT amount (if present)
          example: 1.78
        total:
          type: number
          description: Grand total amount
          example: 34.15
        currency:
          type: string
          description: Currency code inferred from the document
          example: PLN
    OcrStructuredResponseDto:
      type: object
      properties:
        cleanedText:
          type: string
          description: Cleaned, human-readable text reconstructed from OCR
          example: |-
            ZETO Serwis
            ul. Oleska 7
            45-052 Opole
            NIP: 7453009977
            ...
        header:
          description: Document header fields
          allOf:
            - $ref: '#/components/schemas/OcrStructuredHeaderDto'
        items:
          description: Detected line items
          example:
            - name: Ser żółty
              quantity: 0.5
              unit: kg
              unitPrice: 11.9
              lineTotal: 5.95
              vatRate: 8
              currency: PLN
            - name: Jabłka
              quantity: 3
              unitPrice: 4.4
              lineTotal: 13.2
              currency: PLN
          type: array
          items:
            $ref: '#/components/schemas/OcrStructuredItemDto'
        totals:
          description: Totals section
          example:
            subtotal: 32.37
            vatTotal: 1.78
            total: 34.15
            currency: PLN
          allOf:
            - $ref: '#/components/schemas/OcrStructuredTotalsDto'
        warnings:
          description: Non-fatal warnings the model produced while structuring
          example:
            - Uncertain currency
            - Ambiguous date format
          type: array
          items:
            type: string
      required:
        - cleanedText
        - header
        - items
        - totals
...